olist_analytics/requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Machine Learning
scikit-learn>=1.3.0
lifetimes>=0.11.1  # For CLV modeling

# Visualization & Dashboard
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
dash>=2.10.0  # For interactive dashboards
dash-table>=5.0.0  # For interactive tables
dash-bootstrap-components>=1.4.0  # For Bootstrap styling

# Data Quality & Testing
great-expectations>=0.17.0
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0  # For CLI tools
tqdm>=4.65.0  # Progress bars

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
nbformat>=5.9.0

# Workflow Orchestration (optional)
# apache-airflow>=2.6.0  # Uncomment if using Airflow

# API Development (optional)
# fastapi>=0.100.0  # Uncomment if building API
# uvicorn>=0.22.0  # Uncomment if building API